refactor users module with add edit profile
